+++ src/main/java/com/ppnt/client/Packet.java
@@ -3,21 +3,21 @@ package com.ppnt.client;
 import java.io.Serializable;
 
 public class Packet implements Serializable {
-    private int id;
+    private long id;
     private int delay;
     private long timestamp;
 
-    public Packet(int id, int delay, long timestamp) {
+    public Packet(long id, int delay, long timestamp) {
         this.id = id;
         this.delay = delay;
         this.timestamp = timestamp;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
